debt - move some code out of `platform`
