+++ src/vs/base/browser/ui/severityIcon/media/severityIcon.css
