+++ src/vs/workbench/contrib/search/browser/searchMessage.ts
@@ -10,7 +10,7 @@ import { parseLinkedText } from '../../../../base/common/linkedText.js';
 import Severity from '../../../../base/common/severity.js';
 import { IInstantiationService } from '../../../../platform/instantiation/common/instantiation.js';
 import { INotificationService } from '../../../../platform/notification/common/notification.js';
-import { SeverityIcon } from '../../../../platform/severityIcon/browser/severityIcon.js';
+import { SeverityIcon } from '../../../../base/browser/ui/severityIcon/severityIcon.js';
 import { TextSearchCompleteMessage, TextSearchCompleteMessageType } from '../../../services/search/common/searchExtTypes.js';
 import { IOpenerService } from '../../../../platform/opener/common/opener.js';
 import { Schemas } from '../../../../base/common/network.js';
